# Smart UI Element Detection & Zoom System