--- a/src/renderer/modules/config.js
+++ b/src/renderer/modules/config.js
@@ -24,6 +24,14 @@ const config = {
     includeMicrophone: false,
     quality: 'medium',
     framerate: 30,
+  },
+  debug: {
+    showPerformance: false,      // Show performance overlay in preview
+    showUIDetections: true,      // Show UI detection overlays
+    logPerformance: false,       // Log performance metrics to console
+    maxDetectionTime: 50,        // Max time allowed for UI detection (ms)
+    cacheTimeout: 500,           // Cache validity time (ms)
+    degradationThreshold: 30,    // Performance degradation threshold (ms)
   }
 };