--- a/src/renderer/modules/effects.js
+++ b/src/renderer/modules/effects.js
@@ -114,7 +114,160 @@ function renderZoomIndicators({ ctx, canvas, mousePos, velocity }) {
   ctx.fill();
 }
 
+// Render UI detection results in preview
+function renderUIDetectionOverlay({ ctx, canvas, uiDetections, recordingBounds }) {
+  if (!uiDetections || !recordingBounds) return;
+
+  const { formFieldDetections, buttonDetections, currentUIElement } = uiDetections;
+  
+  // Render text field heat map
+  if (formFieldDetections && formFieldDetections.size > 0) {
+    for (const [posKey, detection] of formFieldDetections.entries()) {
+      const [x, y] = posKey.split(',').map(Number);
+      
+      // Convert absolute coordinates to canvas coordinates
+      const relativeX = (x - recordingBounds.x) / recordingBounds.width;
+      const relativeY = (y - recordingBounds.y) / recordingBounds.height;
+      const canvasX = relativeX * canvas.width;
+      const canvasY = relativeY * canvas.height;
+      
+      // Skip if outside canvas bounds
+      if (canvasX < 0 || canvasX > canvas.width || canvasY < 0 || canvasY > canvas.height) {
+        continue;
+      }
+      
+      // Draw text field detection
+      const confidence = detection.avgConfidence || 0.5;
+      const alpha = Math.min(0.6, confidence);
+      
+      ctx.beginPath();
+      ctx.fillStyle = `rgba(100, 149, 237, ${alpha})`; // Cornflower blue
+      ctx.roundRect(canvasX - 30, canvasY - 8, 60, 16, 3);
+      ctx.fill();
+      
+      // Draw confidence indicator
+      ctx.beginPath();
+      ctx.fillStyle = `rgba(255, 255, 255, ${alpha + 0.2})`;
+      ctx.font = '10px Arial';
+      ctx.textAlign = 'center';
+      ctx.fillText(`📝 ${Math.round(confidence * 100)}%`, canvasX, canvasY + 3);
+    }
+  }
+  
+  // Render button heat map
+  if (buttonDetections && buttonDetections.size > 0) {
+    for (const [posKey, detection] of buttonDetections.entries()) {
+      const [x, y] = posKey.split(',').map(Number);
+      
+      // Convert absolute coordinates to canvas coordinates
+      const relativeX = (x - recordingBounds.x) / recordingBounds.width;
+      const relativeY = (y - recordingBounds.y) / recordingBounds.height;
+      const canvasX = relativeX * canvas.width;
+      const canvasY = relativeY * canvas.height;
+      
+      // Skip if outside canvas bounds
+      if (canvasX < 0 || canvasX > canvas.width || canvasY < 0 || canvasY > canvas.height) {
+        continue;
+      }
+      
+      // Draw button detection
+      const confidence = detection.avgConfidence || 0.5;
+      const alpha = Math.min(0.6, confidence);
+      
+      ctx.beginPath();
+      ctx.fillStyle = `rgba(255, 165, 0, ${alpha})`; // Orange
+      ctx.roundRect(canvasX - 20, canvasY - 10, 40, 20, 5);
+      ctx.fill();
+      
+      // Draw confidence indicator
+      ctx.beginPath();
+      ctx.fillStyle = `rgba(255, 255, 255, ${alpha + 0.2})`;
+      ctx.font = '10px Arial';
+      ctx.textAlign = 'center';
+      ctx.fillText(`🔘 ${Math.round(confidence * 100)}%`, canvasX, canvasY + 3);
+    }
+  }
+  
+  // Highlight current active UI element
+  if (currentUIElement && currentUIElement.detection && currentUIElement.detection.position) {
+    const pos = currentUIElement.detection.position;
+    const relativeX = (pos.x - recordingBounds.x) / recordingBounds.width;
+    const relativeY = (pos.y - recordingBounds.y) / recordingBounds.height;
+    const canvasX = relativeX * canvas.width;
+    const canvasY = relativeY * canvas.height;
+    
+    if (canvasX >= 0 && canvasX <= canvas.width && canvasY >= 0 && canvasY <= canvas.height) {
+      // Draw active element highlight
+      ctx.beginPath();
+      ctx.strokeStyle = '#FFD700'; // Gold
+      ctx.lineWidth = 3;
+      ctx.setLineDash([5, 5]);
+      
+      let width = 60, height = 20;
+      if (currentUIElement.type === 'text-field') {
+        width = 80;
+        height = 20;
+      } else if (currentUIElement.type === 'button') {
+        width = 50;
+        height = 25;
+      }
+      
+      ctx.roundRect(canvasX - width/2, canvasY - height/2, width, height, 5);
+      ctx.stroke();
+      ctx.setLineDash([]);
+      
+      // Draw element type label
+      ctx.beginPath();
+      ctx.fillStyle = 'rgba(255, 215, 0, 0.9)'; // Gold background
+      ctx.roundRect(canvasX - 25, canvasY - 35, 50, 15, 3);
+      ctx.fill();
+      
+      ctx.fillStyle = '#000';
+      ctx.font = 'bold 10px Arial';
+      ctx.textAlign = 'center';
+      const icon = currentUIElement.type === 'text-field' ? '📝' : 
+                   currentUIElement.type === 'button' ? '🔘' : '🎯';
+      ctx.fillText(`${icon} ${currentUIElement.type}`, canvasX, canvasY - 25);
+    }
+  }
+}
+
+// Render performance metrics overlay
+function renderPerformanceOverlay({ ctx, canvas, performanceData }) {
+  if (!performanceData) return;
+  
+  const { detectionFPS, avgDetectionTime, cacheHitRate, totalDetections } = performanceData;
+  
+  // Draw performance panel in top-right corner
+  const panelWidth = 180;
+  const panelHeight = 80;
+  const panelX = canvas.width - panelWidth - 10;
+  const panelY = 10;
+  
+  // Background
+  ctx.beginPath();
+  ctx.fillStyle = 'rgba(0, 0, 0, 0.7)';
+  ctx.roundRect(panelX, panelY, panelWidth, panelHeight, 5);
+  ctx.fill();
+  
+  // Text
+  ctx.fillStyle = '#FFF';
+  ctx.font = '11px Arial';
+  ctx.textAlign = 'left';
+  
+  let textY = panelY + 15;
+  ctx.fillText(`🎯 UI Detection: ${detectionFPS || 0} fps`, panelX + 5, textY);
+  textY += 15;
+  ctx.fillText(`⏱️ Avg Time: ${avgDetectionTime || 0}ms`, panelX + 5, textY);
+  textY += 15;
+  ctx.fillText(`💾 Cache Hit: ${Math.round((cacheHitRate || 0) * 100)}%`, panelX + 5, textY);
+  textY += 15;
+  ctx.fillText(`📊 Total: ${totalDetections || 0} detections`, panelX + 5, textY);
+}
+
 module.exports = {
   renderMouseEffects,
-  renderZoomIndicators
+  renderZoomIndicators,
+  renderUIDetectionOverlay,
+  renderPerformanceOverlay
 };