… PR_DESCRIPTION.md
… reference_code.md
… src/renderer/modules/preview.js
… src/renderer/modules/streamProcessor.js
… src/renderer/modules/uiDetection.js
… tasks.md